requirements.txt
ultralytics>=8.0,<9.0
deep-sort-realtime>=1.3,<2.0
mysql-connector-python>=8.0,<9.0
opencv-python>=4.8,<5.0
numpy>=1.24,<2.0
torch>=2.0,<3.0
